+++ src/main/java/org/mbari/mxm/provider_client/responses/MissionStatusResponse.java
@@ -5,12 +5,13 @@ import lombok.Data;
 // TODO this is all preliminary
 
 @Data
-public class PostMissionResponse {
+public class MissionStatusResponse {
 
   public PMResponse result;
 
   @Data
   public static class PMResponse {
+    public String missionId; // captured as `providerMissionId` in MXM mission model
     public String status;
   }
 }